added main game loop
