--- Falcon/FALCON_GAME/headers/GameState.h
@@ -1,11 +1,0 @@
-#ifndef GAME_STATE_H
-#define GAME_STATE_H
-
-class GameState 
-{
-	virtual void Update();
-	virtual ~GameState();
-};
-
-#endif // !1
-
